requirements.txt
pyserial~=3.5
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.4
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
zipp==3.1.0